文章分页，浏览量

+++ article/views.py
@@ -9,15 +9,31 @@ from .forms import ArticlePostForm
 # 引入User模型
 from django.contrib.auth.models import User
 from django.contrib.auth.decorators import login_required
+from django.core.paginator import Paginator
 
 
 # 文章列表
 def article_list(request):
 
-    # 取出所有的博客文章
-    articles = ArticlePost.objects.all()
+    # 根据GET请求中查询条件
+    # 返回不同的排序对象数组
+    if request.GET.get('order') == 'total_views':
+        article_list = ArticlePost.objects.all().order_by('-total_views')
+        order = 'total_views'
+    else:
+        article_list = ArticlePost.objects.all()
+        order = 'normal'
+
+    # 每页显示1篇文章
+    paginator = Paginator(article_list, 3)
+
+    # 获取url中的页面
+    page = request.GET.get('page')
+
+    # 将导航栏对象相应的页码内容返回给 articles
+    articles = paginator.get_page(page)
     # 需要传递给模板（templates）的对象
-    context = {'articles': articles}
+    context = {'articles': articles, 'order': order}
     # render函数，载入模板并返回context对象
     return render(request, 'article/list.html', context)
 
@@ -26,6 +42,10 @@ def article_list(request):
 def article_detail(request, pk):
     # 取出所有文章
     article = ArticlePost.objects.get(pk=pk)
+
+    # 浏览量+ 1
+    article.total_views += 1
+    article.save(update_fields=['total_views'])
 
     # 将Markdown语法渲染成HTML的样式
     article.body = markdown.markdown(article.body,
@@ -77,9 +97,13 @@ def article_create(request):
 def article_delete(request, pk):
     # 根据对应的id去删除对应的文章
     article = ArticlePost.objects.get(pk=pk)
-    # 调用。delete()方法
-    article.delete()
-    return redirect('article:article_list')
+    # 过滤费作者用户
+    if request.user != article.author:
+        return HttpResponse('抱歉，你无权修改这篇文章')
+    else:
+        # 调用。delete()方法
+        article.delete()
+        return redirect('article:article_list')
 
 
 # 编辑文章
@@ -95,29 +119,35 @@ def article_update(request, pk):
 
     # 获取需要修改的具体文章对象
     article = ArticlePost.objects.get(pk=pk)
-    # 判断用户是否为POST表单提交数据
-    if request.method == 'POST':
-        # 将提交数据赋值到表单实例中
-        article_post_form = ArticlePostForm(data=request.POST)
-        # 判断提交的数据是否满足模型要求
-        if article_post_form.is_valid():
-            # 保存新写入的title，body数据并保存
-            article.title = request.POST['title']
-            article.body = request.POST['body']
-            article.save()
-            # 完成之后返回到修改后的文章中，需要传入文章的id值
-            return redirect('article:article_detail', pk=pk)
-        # 如果不合法，返回错误信息
-        else:
-            return HttpResponse('表单数据内容有误，请重新填写。')
 
-    # 如果用户是GET请求获取数据
+    # 过滤费作者用户
+    if request.user != article.author:
+        return HttpResponse('抱歉，你无权修改这篇文章')
     else:
 
-        # 创建表单实例
-        article_post_form = ArticlePostForm()
-        # 赋值上下文。将article文章对象也传递进去，以便提取旧的内容
-        context = {'article': article, 'article_post_form':article_post_form}
-        # 将响应返回到模板中去
-        return render(request, 'article/update.html', context)
+        # 判断用户是否为POST表单提交数据
+        if request.method == 'POST':
+            # 将提交数据赋值到表单实例中
+            article_post_form = ArticlePostForm(data=request.POST)
+            # 判断提交的数据是否满足模型要求
+            if article_post_form.is_valid():
+                # 保存新写入的title，body数据并保存
+                article.title = request.POST['title']
+                article.body = request.POST['body']
+                article.save()
+                # 完成之后返回到修改后的文章中，需要传入文章的id值
+                return redirect('article:article_detail', pk=pk)
+            # 如果不合法，返回错误信息
+            else:
+                return HttpResponse('表单数据内容有误，请重新填写。')
+
+        # 如果用户是GET请求获取数据
+        else:
+
+            # 创建表单实例
+            article_post_form = ArticlePostForm()
+            # 赋值上下文。将article文章对象也传递进去，以便提取旧的内容
+            context = {'article': article, 'article_post_form':article_post_form}
+            # 将响应返回到模板中去
+            return render(request, 'article/update.html', context)
